+++ Assets/Scripts/Game/Characters/Attacks/MeleeAttack.cs
@@ -4,7 +4,7 @@ namespace Game.Characters.Attacks
 {
     public class MeleeAttack: IAttack
     {
-        public void Attack(int damage, Health targetHP)
+        public void Attack(float damage, Health targetHP)
         {
             targetHP.GetDamage(damage);
         }
